Added a Skybox demo to check how to create it.
Fixed the Teleport Validity Rules: That's is how to do that:
https://learn.unity.com/tutorial/unit-4-locomotion-and-ergonomics?courseId=5d955b5dedbc2a319caab9a0#5d9691a1edbc2a2d9fcf776e

Add a new layer named “Teleportable” and set the FloorCollider object to the new layer.
Create a new empty GameObject and  use the "Add Component" button in the Inspector to attach the AnyLayerRule script.
Set the LayerMask field in the AnyLayerRule script to the Teleportable layer.
Set the TargetValidity field in the TeleporterFacade script to AnyLayerRule script gameobject.
For each ObjectPointerCurved assign the AnyLayerRule to the TargetValidity field of the PointerFacade script.

LookAt() works (you have to enable the ApplyRotation, see code)